# swagger/RPX100/api/swagger/swagger.yaml
swagger: "2.0"
info:
  description: >
    This is the web REST/JSON API of RPX100 SDR software. RPX100 is an Open Source Qt5/OpenGL 3.0+ (4.3+ in Windows) GUI and server Software Defined Radio and signal analyzer in software. It supports Airspy, BladeRF, HackRF, LimeSDR, PlutoSDR, RTL-SDR, SDRplay RSP1 and FunCube

      ---
      Limitations and specifcities:

      * In RPX100 GUI the first Rx device set cannot be deleted. Conversely the server starts with no device sets and its number of device sets can be reduced to zero by as many calls as necessary to /RPX100/deviceset with DELETE method.
      * Preset import and export from/to file is a server only feature.
      * Device set focus is a GUI only feature.
      * The following channels are not implemented (status 501 is returned): ATV and DATV demodulators, Channel Analyzer NG, LoRa demodulator
      * The device settings and report structures contains only the sub-structure corresponding to the device type. The DeviceSettings and DeviceReport structures documented here shows all of them but only one will be or should be present at a time
      * The channel settings and report structures contains only the sub-structure corresponding to the channel type. The ChannelSettings and ChannelReport structures documented here shows all of them but only one will be or should be present at a time

      ---

  version: "6.0.0"
  title: RPX100
  contact:
    url: "https://github.com/f4exb/RPX100"
    email: "f4exb06@gmail.com"
...
basePath: /
...
schemes:
...
consumes:
...
produces:
...
paths:
  /RPX100:
    # binds a127 app logic to a route
    x-swagger-router-controller: instance
    get:
      description: RPX100 instance summary
      # used as the method name of the controller
      operationId: instanceSummary
      tags:
        - Instance
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/InstanceSummaryResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
    delete:
      description: Stop RPX100 instance (server only)
      # used as the method name of the controller
      operationId: instanceDelete
      tags:
        - Instance
      responses:
        "202":
          description: Message to stop the RPX100 instance was sent successfully
          schema:
            # a pointer to a definition
            $ref: "#/definitions/InstanceSummaryResponse"
        "400":
          description: Unsupported (GUI instance)
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
...
  /RPX100/deviceset/{deviceSetIndex}/focus:
    x-swagger-router-controller: deviceset
    patch:
      description: Set focus on device set (GUI only)
      operationId: devicesetFocusPatch
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
      responses:
        "202":
          description: Message to focus on device set was sent successfully
          schema:
            $ref: "#/definitions/SuccessResponse"
        "400":
          description: Unsupported (srver instance)
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Invalid index
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
...
  /RPX100/deviceset/{deviceSetIndex}/device/settings:
    x-swagger-router-controller: deviceset
    get:
      description: Get device settings
      operationId: devicesetDeviceSettingsGet
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
      responses:
        "200":
          description: On success returns current settings values
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
        "404":
          description: Invalid device set index or device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
    put:
      description: Apply all settings unconditionally (force)
      operationId: devicesetDeviceSettingsPut
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
        - name: body
          in: body
          description: Device settings to apply
          required: true
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
      responses:
        "200":
          description: On success returns new settings values
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
        "404":
          description: Invalid device set index or device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
    patch:
      description: Apply settings differentially (no force)
      operationId: devicesetDeviceSettingsPatch
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
        - name: body
          in: body
          description: Device settings to apply
          required: true
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
      responses:
        "200":
          description: On success returns new settings values
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
        "404":
          description: Invalid device set index or device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
  /RPX100/deviceset/{deviceSetIndex}/device/run:
    x-swagger-router-controller: deviceset
    get:
      description: get device run status for a single subsystem device (Rx or Tx)
      operationId: devicesetDeviceRunGet
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
      responses:
        "200":
          description: On success return current state
          schema:
            $ref: "#/definitions/DeviceState"
        "400":
          description: Invalid device set index
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
    post:
      description: start device for a single subsystem device (Rx or Tx)
      operationId: devicesetDeviceRunPost
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
        - name: body
          in: body
          description: Originator information in the reverse API case
          required: false
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
      responses:
        "200":
          description: On success return state before change
          schema:
            $ref: "#/definitions/DeviceState"
        "400":
          description: Invalid device set index
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
    delete:
      description: stop device for a single subsystem device (Rx or Tx)
      operationId: devicesetDeviceRunDelete
      tags:
        - DeviceSet
      parameters:
        - in: path
          name: deviceSetIndex
          type: integer
          required: true
          description: Index of device set in the device set list
        - name: body
          in: body
          description: Originator information in the reverse API case
          required: false
          schema:
            $ref: "http://swgserver:8081/api/swagger/include/DeviceSettings.yaml#/DeviceSettings"
      responses:
        "200":
          description: On success return state before change
          schema:
            $ref: "#/definitions/DeviceState"
        "400":
          description: Invalid device set index
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Device not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/Response_500"
        "501":
          $ref: "#/responses/Response_501"
...
definitions:
  InstanceSummaryResponse:
    description: "Summarized information about this RPX100 instance"
    required:
      - version
      - qtVersion
      - dspRxBits
      - dspTxBits
      - pid
      - appname
      - devicesetlist
      - architecture
      - os
    properties:
      version:
        description: "Current software version"
        type: string
      qtVersion:
        description: "Qt version with which the software was compiled"
        type: string
      dspRxBits:
        description: "Number of samples significant bits in software Rx DSP"
        type: integer
      dspTxBits:
        description: "Number of samples significant bits in software Tx DSP"
        type: integer
      pid:
        description: "PID of the RPX100 instance"
        type: integer
      appname:
        description: "Application name: RPX100 for a GUI instance and RPX100Srv for a server instance"
        type: string
      architecture:
        description: "Codename of the CPU architecture on which the instance is running (available with Qt >= 5.4)"
        type: string
      os:
        description: "Descriptive text of the operating system running the instance (available with Qt >= 5.4)"
        type: string
      logging:
        $ref: "#/definitions/LoggingInfo"
      devicesetlist:
        $ref: "#/definitions/DeviceSetList"
      featuresetlist:
        $ref: "#/definitions/FeatureSetList"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
        example: "KO"
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  LoggingInfo:
    description: "Logging parameters setting"
    properties:
      consoleLevel:
        description: "Minimum level of messages printed to console: debug, info, warning, error"
        type: string
      fileLevel:
        description: "Minimum level of messages written to file: debug, info, warning, error"
        type: string
      dumpToFile:
        description: "not zero (true) if messages are written to file"
        type: integer
      fileName:
        description: "Name of the log file"
        type: string
...
  DeviceSet:
    description: "Sampling device and its associated channels"
    required:
      - samplingDevice
      - channelcount
    properties:
      samplingDevice:
        $ref: "#/definitions/SamplingDevice"
      channelcount:
        description: "Number of channels in the set"
        type: integer
      channels:
        description: "Channels list"
        type: array
        items:
          $ref:  "#/definitions/Channel"
  DeviceSetList:
    description: "List of device sets opened in this instance"
    required:
      - devicesetcount
    properties:
      devicesetcount:
        description: "Number of device sets opened in this instance"
        type: integer
      devicesetfocus:
        description: "Index of the deviceset that has focus (relevant for GUI only)"
        type: integer
      deviceSets:
        type: array
        items:
          $ref: "#/definitions/DeviceSet"
  Feature:
    description: "Feature summarized information"
    required:
      - index
      - id
      - uid
      - title
    properties:
      index:
        description: "Index in the list of features"
        type: integer
      id:
        description: "Key to identify the type of feature"
        type: string
      uid:
        description: "Feature instance unique id"
        type: integer
        format: int64
      title:
        description: "Freeform title of the feature"
        type: string
  FeatureSet:
    description: "Grouping of features"
    required:
      - index
      - featurecount
    properties:
      index:
        description: "Index in the list of feature sets opened in this instance"
        type: integer
      featurecount:
        description: "Number of features in the set"
        type: integer
      features:
        description: "Features list"
        type: array
        items:
          $ref:  "#/definitions/Feature"
  FeatureSetList:
    description: "List of feature sets opened in this instance"
    required:
      - featuresetcount
    properties:
      featuresetcount:
        description: "Number of feature sets opened in this instance"
        type: integer
      featureSets:
        type: array
        items:
          $ref: "#/definitions/FeatureSet"
...
  DeviceState:
    description: "Device running state"
    properties:
      state:
        description: "Unique state: notStarted, idle, ready, running, error"
        type: string
  SamplingDevice:
    description: "Information about a logical device available from an attached hardware device that can be used as a sampling device"
    required:
      - index
      - hwType
      - streamIndex
      - sequence
      - serial
      - centerFrequency
      - bandwidth
    properties:
      index:
        description: "Index in the list of device sets opened in this instance"
        type: integer
      hwType:
        description: "Key to identify the type of hardware device"
        type: string
      direction:
        description: "0 for Rx only, 1 for Tx only or 2 for any number and direction (default 0)"
        type: integer
      deviceNbStreams:
        description: "Number of channels or streams in the device"
        type: integer
      deviceStreamIndex:
        description: "Index of the channel in the device. Applies to single stream directions (Rx or Tx)"
        type: integer
      sequence:
        description: "Sequence in the enumeration of same device types"
        type: integer
      serial:
        description: "Serial number of the hardware device"
        type: string
      centerFrequency:
        description: "Center frequency in Hz at which the device is tuned possibly taking NCO into account"
        type: integer
        format: int64
      bandwidth:
        description: "Bandwidth in Hz or complex S/s of baseband"
        type: integer
      state:
        description: "Single subsystem state: notStarted, idle, ready, running, error"
        type: string
      stateRx:
        description: "Rx subsystem state (MIMO): notStarted, idle, ready, running, error"
        type: string
      stateTx:
        description: "Tx subsystem state (MIMO): notStarted, idle, ready, running, error"
        type: string
  Channel:
    description: "Channel summarized information"
    required:
      - index
      - id
      - uid
      - title
      - deltaFrequency
    properties:
      direction:
        description: 0 for Rx only, 1 for Tx only or 2 for any direction (MIMO)
        type: integer
      index:
        description: "Index in the list of channels"
        type: integer
      id:
        description: "Key to identify the type of channel"
        type: string
      uid:
        description: "Channel instance unique id"
        type: integer
        format: int64
      title:
        description: "Freeform title of the channel"
        type: string
      deltaFrequency:
        description: "Channel shift frequency in Hz from the center of baseband"
        type: integer
      report:
        $ref: "http://swgserver:8081/api/swagger/include/ChannelReport.yaml#/ChannelReport"
...
responses:
  Response_500:
    description: Error
    schema:
      $ref: "#/definitions/ErrorResponse"
  Response_501:
    description: Function not implemented
    schema:
      $ref: "#/definitions/ErrorResponse"